New domain Web (rails engine) with the following features extracted from the main app: - cookies banner - cookies policy page - cookies policy and privacy policy links in the footer
