+++ .rubocop.yml
@@ -16,6 +16,8 @@ AllCops:
     - !ruby/regexp /old_and_unused\.rb$/
     # The parser gem fails to parse this file with out current Ruby version.
     - 'spec/factories.rb'
+    # Excluding: inadequate Naming/FileName rule rejects GemFile name with camelcase 
+    - 'engines/web/Gemfile'
 
 # OFN SETTINGS
 # Cop settings that have been agreed upon by the OFN community
