+++ app/views/layouts/darkswarm.html.haml
@@ -19,6 +19,8 @@
     %script{src: "//maps.googleapis.com/maps/api/js?libraries=places,geometry#{ ENV['GOOGLE_MAPS_API_KEY'] ? '&key=' + ENV['GOOGLE_MAPS_API_KEY'] : ''} "}
     = split_stylesheet_link_tag "darkswarm/all"
     = javascript_include_tag "darkswarm/all"
+    = split_stylesheet_link_tag "web/all"
+    = javascript_include_tag "web/all"
 
     = render "layouts/i18n_script"
     = render "layouts/bugherd_script"
