+++ config/routes.rb
@@ -105,16 +105,15 @@ Openfoodnetwork::Application.routes.draw do
       get :job_queue
     end
 
-    scope '/cookies' do
-      resource :consent, only: [:show, :create, :destroy], :controller => "cookies_consent"
-    end
-
     resources :customers, only: [:index, :update]
 
     post '/product_images/:product_id', to: 'product_images#update_product_image'
   end
 
   get 'sitemap.xml', to: 'sitemap#index', defaults: { format: 'xml' }
+
+  # Mount Web engine routes
+  mount Web::Engine, :at => '/'
 
   # Mount Spree's routes
   mount Spree::Core::Engine, :at => '/'
