+++ engines/web/README.rdoc
@@ -1,0 +1,3 @@
+= Web
+
+This is the rails engine for the WEB OFN domain.
