+++ engines/web/app/assets/javascripts/web/cookies_banner/cookies_banner_controller.js.coffee
