+++ engines/web/app/assets/javascripts/web/cookies_policy/cookies_policy_modal_directive.js.coffee
