+++ engines/web/app/assets/stylesheets/web/cookies_banner.css.scss
@@ -1,4 +1,4 @@
-@import 'branding';
+@import '../../../../../../app/assets/stylesheets/darkswarm/branding';
 
 .cookies-banner {
   background: $dark-grey;
