+++ engines/web/app/assets/stylesheets/web/cookies_policy_modal.css.scss
@@ -1,4 +1,4 @@
-@import 'branding';
+@import '../../../../../../app/assets/stylesheets/darkswarm/branding';
 
 .cookies-policy-modal {
   background: $disabled-light;
