+++ engines/web/app/controllers/web/application_controller.rb
@@ -1,0 +1,5 @@
+module Web
+  class ApplicationController < ActionController::Base
+    protect_from_forgery with: :exception
+  end
+end
