+++ engines/web/app/views/angular_templates/cookies_banner.html.haml
