+++ engines/web/lib/web.rb
@@ -1,0 +1,4 @@
+require "web/engine"
+
+module Web
+end
