+++ engines/web/lib/web/engine.rb
@@ -1,0 +1,5 @@
+module Web
+  class Engine < ::Rails::Engine
+    isolate_namespace Web
+  end
+end
