+++ engines/web/lib/web/version.rb
@@ -1,0 +1,3 @@
+module Web
+  VERSION = "0.0.1".freeze
+end
