+++ engines/web/web.gemspec
@@ -1,0 +1,13 @@
+$LOAD_PATH.push File.expand_path('lib', __dir__)
+
+require "web/version"
+
+Gem::Specification.new do |s|
+  s.name        = "web"
+  s.version     = Web::VERSION
+  s.authors     = ["developers@ofn"]
+  s.summary     = "Web domain of the OFN solution."
+
+  s.files = Dir["{app,config,db,lib}/**/*"] + ["LICENSE.txt", "Rakefile", "README.rdoc"]
+  s.test_files = Dir["test/**/*"]
+end
